requirements.txt
discord.py 
openai 
google-generativeai 
langchain 
faiss-cpu 
python-dotenv 
youtube_dl 
requests